+++ crates/cli/src/commands/completions.rs
@@ -1,7 +1,7 @@
 use crate::app::App;
 use clap::CommandFactory;
 use clap_complete::{generate, Shell};
-use proto::ProtoError;
+use proto_core::ProtoError;
 
 pub async fn completions(shell: Option<Shell>) -> Result<(), ProtoError> {
     let Some(shell) = shell.or_else(Shell::from_env) else {
